api/requirements.txt
boto3
pandas
PyMuPDF
faiss-cpu
numpy
requests
flask
python-docx
llama-index
llama-index-embeddings-gemini
flask_cors